requirements.txt
openai>=1.0.0
pyyaml>=6.0
streamlit>=1.30.0
pymysql>=1.1.0
